third_party/blink/web_tests/media/W3C/video/readyState/readyState_during_canplaythrough-expected.txt

video.readyState should be HAVE_ENOUGH_DATA during canplaythrough event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS "4" is "4"
PASS successfullyParsed is true

TEST COMPLETE


spec reference
